move reviews to slot
