+++ blocks/product-details/product-details.css
@@ -1,1 +1,1 @@
-/* stylelint-disable no-empty-source */
+@import 'product.css';
