+++ blocks/product-details/product-details.js
@@ -10,10 +10,12 @@ import ProductDetails from '@dropins/storefront-pdp/containers/ProductDetails.js
 // Libs
 import { getProduct, getSkuFromUrl, setJsonLd } from '../../scripts/commerce.js';
 import { getConfigValue } from '../../scripts/configs.js';
-import { fetchPlaceholders, loadScript } from '../../scripts/aem.js';
+import { fetchPlaceholders } from '../../scripts/aem.js';
 import { pushViewItemEvent } from '../../scripts/gtm.js';
+
+// Slots
 import Title from './slots/Title.js';
-import addSkeleton from '../../scripts/components/skeleton/skeleton.js';
+import Content from './slots/Content.js';
 
 // Error Handling (404)
 async function errorGettingProduct(code = 404) {
@@ -108,35 +110,6 @@ function setMetaTags(product) {
   createMetaTag('twitter:card', product.shortDescription, 'name');
   createMetaTag('twitter:title', product.metaTitle, 'name');
   createMetaTag('twitter:image', metaImage, 'name');
-}
-
-async function renderReviewsOnVisible(block) {
-  const wrapper = document.createElement('div');
-  wrapper.className = 'yotpo yotpo-main-widget';
-  wrapper.setAttribute('data-name', 'Endeavor Daytrip Backpack');
-  wrapper.setAttribute('data-product-id', '2250');
-  wrapper.setAttribute('data-product-id', '2250');
-  wrapper.setAttribute('data-url', 'spacedye-hi-lo-tank-updated');
-  wrapper.setAttribute('data-description', '');
-  wrapper.setAttribute('data-yotpo-element-id', '1');
-  wrapper.setAttribute('data-image-url', 'http://integration-5ojmyuq-7yvbzwvtkgerq.us-4.magentosite.cloud/media/catalog/product/w/b/wb06-red-0.jpg?auto=webp&quality=80&crop=false&fit=cover&width=960');
-  addSkeleton(wrapper, {
-    height: 400,
-    width: '100%',
-  });
-  block.appendChild(wrapper);
-
-  const apiKey = await getConfigValue('yotpo-api-key');
-  const observer = new IntersectionObserver((entries, obs) => {
-    entries.forEach((entry) => {
-      if (entry.isIntersecting && !window?.yotpo?.initialized) {
-        loadScript(`//staticw2.yotpo.com/${apiKey}/widget.js`, { id: 'yotpo-script-loader', async: true, defer: true });
-        obs.unobserve(entry.target);
-      }
-    });
-  });
-
-  observer.observe(wrapper);
 }
 
 export default async function decorate(block) {
@@ -247,6 +220,7 @@ export default async function decorate(block) {
           },
           slots: {
             Title: (ctx) => Title(ctx, block),
+            Content: (ctx) => Content(ctx, block),
             Actions: (ctx) => {
               // Add to Cart Button
               ctx.appendButton((next, state) => {
@@ -281,8 +255,6 @@ export default async function decorate(block) {
           },
           useACDL: true,
         })(block);
-        // add reviews after containers are rendered
-        renderReviewsOnVisible(block);
       } catch (e) {
         console.error(e);
         await errorGettingProduct();
